+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Delete: allow versioned keys
 - delete: check if dir exists before call trash to avoid process freeze
 - validation: main file validation checks for text content not only extensions
+- tests: added an extra remote profile on delete registered module test #235
 
 ### Added
 - Validation: add check for open format extension